modules/tg-export/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-export"
version = "1.0.0"
description = "Экспорт и анализ Telegram чатов, каналов и личных переписок"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "telethon>=1.34.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
rich = ["rich>=13.0"]

[project.scripts]
tg-export = "tg_export.cli:main"

[tool.setuptools.packages.find]
include = ["tg_export*"]
